+++ QuestAnswerApi/Repositories/CardRepositories/ICardRepository.cs
@@ -3,9 +3,9 @@ namespace QuestAnswerApi.Repositories.CardRepositories;
 using QuestAnswerApi.Models;
 public interface ICardRepository
 {
-    Task<Card?> GetCardByIdAsync(int id);
+    Task<Card?> GetCardByIdAsync(long id);
     Task<IEnumerable<Card>> GetAllCardsAsync();
     Task CreateCardAsync(Card card);
     Task<bool> UpdateCardAsync(Card card);
-    Task<bool> DeleteCardAsync(int id);
+    Task<bool> DeleteCardAsync(long id);
 }
